Pendual: Finish the majority of the events (#44)
* Pendual: Finish the majority of the events

In this commit I have pushed my current changes to IIDX 22 Pendual. I've completed every event other than Qpronicle Chord as I have ran out of steam on this mix for now. I will return and try to 100% this game sooner than later though. For now i'll leave it at this.

I have made the events all Frontend modifiable so that's good enough for now. I've thrown in a ton of #TODO statements to show my current progress.

* Pendual: Remove clutter and add back omni-event control

In this commit I have re-added the option to disable events in omnimix and have remove code that was not needed/not fully finished that I added in the past commit. I've also reverted changes that were not relevant nor useful.